+++ src/commands/gists/insert.ts
@@ -42,7 +42,7 @@ const insert: CommandInitializer = (
     } catch (err) {
       const error: Error = err as Error;
       logger.error(`${command} > ${error && error.message}`);
-      insights.exception(command, { messsage: error.message });
+      insights.exception(command, { message: error.message });
       utils.notify.error('Could Not Insert', `Reason: ${error.message}`);
     }
   };