+++ src/commands/gists/update-access-key.ts
@@ -1,7 +1,7 @@
 import { GistCommands } from '../extension-commands';
 
 const updateAccessKey: CommandInitializer = (
-  _config: Configuration,
+  config: Configuration,
   services: Services,
   utils: Utils
 ): [Command, CommandFn] => {
@@ -12,21 +12,40 @@ const updateAccessKey: CommandInitializer = (
   const commandFn = (): void => {
     try {
       const profile = profiles.get();
-      let gistUrl: string;
+      let gistUrl = 'unknown';
+      let overridesApplied = 'false';
       if (profile) {
-        const key = profile.key;
-        const url = (gistUrl = profile.url);
-        gists.configure({ key, url });
+        let optionOverride: GistServiceOptions = {};
+        const profileOptionOverride = config.get<{
+          [profile: string]: GistServiceOptions;
+        }>('profileOptions');
+
+        if (
+          profileOptionOverride &&
+          Object.keys(profileOptionOverride).length > 0 &&
+          profileOptionOverride[profile.name]
+        ) {
+          overridesApplied = 'true';
+          optionOverride = { ...profileOptionOverride[profile.name] };
+        }
+        const key = optionOverride.key || profile.key;
+        const url = optionOverride.url || (gistUrl = profile.url);
+        const rejectUnauthorized = optionOverride.rejectUnauthorized || true;
+        gists.configure({ key, url, rejectUnauthorized });
       } else {
         gistUrl = 'reset';
-        gists.configure({ key: undefined, url: undefined });
+        gists.configure({
+          key: undefined,
+          rejectUnauthorized: undefined,
+          url: undefined
+        });
       }
-      insights.track(command, { url: gistUrl });
+      insights.track(command, { url: gistUrl, overridesApplied });
       logger.debug('updated access key');
     } catch (err) {
       const error: Error = err as Error;
       logger.error(`${command} > ${error && error.message}`);
-      insights.exception(command, { messsage: error.message });
+      insights.exception(command, { message: error.message });
       utils.notify.error('Could Not Update Access Key', error.message);
     }
   };