+++ src/commands/profiles/select.ts
@@ -59,7 +59,7 @@ const select: CommandInitializer = (
     } catch (err) {
       const error: Error = err as Error;
       logger.error(`${command} > ${error && error.message}`);
-      insights.exception(command, { messsage: error.message });
+      insights.exception(command, { message: error.message });
       utils.notify.error(
         'Could Not Select Profile',
         `Reason: ${error.message}`